requirements.txt.txt
Python IDE, MySQL workbench and the following libraries:
MySQL-connector-python
datetime
tkinter